verificacao_certificacao/requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
requests>=2.31.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
gspread>=6.0.0
